ECE 385 Final Project: Street Fighter FPGA
## Details
Made a 2 player street fighter level between Ryu and Akuma. Player 1 moves with wasd with c to attach. Player 2 moves with ijkl with n to attack.

## How to Run
Add all files under files/ to the Quartus project. Then add all sprites by going to all ryu*, akuma*, and ko folders to add the .qip files to the project. Finally, set fighter as the top level and compile the project. The C code can be added as it is under the software/ folder.